Features: Split MiniMapFeature and MapFeature
