+++ common/src/main/java/com/wynntils/screens/WynntilsMenuScreen.java
@@ -6,6 +6,7 @@ package com.wynntils.screens;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.wynntils.core.webapi.WebManager;
+import com.wynntils.features.user.overlays.map.MapFeature;
 import com.wynntils.mc.objects.CommonColors;
 import com.wynntils.mc.objects.CustomColor;
 import com.wynntils.mc.render.FontRenderer;
@@ -101,21 +102,23 @@ public class WynntilsMenuScreen extends WynntilsMenuScreenBase {
                         new TextComponent(""),
                         new TranslatableComponent("screens.wynntils.wynntilsMenu.leftClickToSelect")
                                 .withStyle(ChatFormatting.GREEN))));
-        buttons.add(new WynntilsMenuButton(
-                Texture.MAP_ICON,
-                true,
-                new MainMapScreen(),
-                List.of(
-                        new TextComponent("[>] ")
-                                .withStyle(ChatFormatting.GOLD)
-                                .append(new TranslatableComponent("screens.wynntils.wynntilsQuestBook.mainMap.name")
-                                        .withStyle(ChatFormatting.BOLD)
-                                        .withStyle(ChatFormatting.GOLD)),
-                        new TranslatableComponent("screens.wynntils.wynntilsQuestBook.mainMap.description")
-                                .withStyle(ChatFormatting.GRAY),
-                        new TextComponent(""),
-                        new TranslatableComponent("screens.wynntils.wynntilsMenu.leftClickToSelect")
-                                .withStyle(ChatFormatting.GREEN))));
+        if (MapFeature.INSTANCE.isEnabled()) {
+            buttons.add(new WynntilsMenuButton(
+                    Texture.MAP_ICON,
+                    true,
+                    new MainMapScreen(),
+                    List.of(
+                            new TextComponent("[>] ")
+                                    .withStyle(ChatFormatting.GOLD)
+                                    .append(new TranslatableComponent("screens.wynntils.wynntilsQuestBook.mainMap.name")
+                                            .withStyle(ChatFormatting.BOLD)
+                                            .withStyle(ChatFormatting.GOLD)),
+                            new TranslatableComponent("screens.wynntils.wynntilsQuestBook.mainMap.description")
+                                    .withStyle(ChatFormatting.GRAY),
+                            new TextComponent(""),
+                            new TranslatableComponent("screens.wynntils.wynntilsMenu.leftClickToSelect")
+                                    .withStyle(ChatFormatting.GREEN))));
+        }
 
         assert buttons.size() <= 8;
     }
