+++ common/src/main/resources/assets/wynntils/lang/en_us.json
@@ -292,26 +292,31 @@
   "feature.wynntils.lootrunUtils.undoStandNear": "Stand near a previous spot to undo movements.",
   "feature.wynntils.lootrunUtils.undoSuccessful": "Successfully undid movements.",
   "feature.wynntils.map.name": "Map",
-  "feature.wynntils.map.overlay.miniMap.borderType.description": "What style should the minimap border have?",
-  "feature.wynntils.map.overlay.miniMap.borderType.name": "Mini Map Border Type",
-  "feature.wynntils.map.overlay.miniMap.description": "Mini Map that displays region surrounding player",
-  "feature.wynntils.map.overlay.miniMap.followPlayerRotation.description": "Should the minimap be locked facing north or rotate based on the direction you're facing?",
-  "feature.wynntils.map.overlay.miniMap.followPlayerRotation.name": "Follow Player Rotation",
-  "feature.wynntils.map.overlay.miniMap.maskType.description": "Should the minimap be a rectangle or a circle? (Circle is unsupported as of now)?",
-  "feature.wynntils.map.overlay.miniMap.maskType.name": "Map Mask Type",
-  "feature.wynntils.map.overlay.miniMap.name": "Mini Map Overlay",
-  "feature.wynntils.map.overlay.miniMap.pointerColor.description": "What colour should the minimap pointer be?",
-  "feature.wynntils.map.overlay.miniMap.pointerColor.name": "Pointer Colour",
-  "feature.wynntils.map.overlay.miniMap.pointerType.description": "What style should the minimap pointer have?",
-  "feature.wynntils.map.overlay.miniMap.pointerType.name": "Mini Map Pointer Type",
-  "feature.wynntils.map.overlay.miniMap.renderUsingLinear.description": "Should the map be rendered using linear textures to avoid aliasing issues?",
-  "feature.wynntils.map.overlay.miniMap.renderUsingLinear.name": "Use Linear Rendering",
-  "feature.wynntils.map.overlay.miniMap.scale.description": "How zoomed in should the map be?",
-  "feature.wynntils.map.overlay.miniMap.scale.name": "Map Scale",
-  "feature.wynntils.map.overlay.miniMap.showCompass.description": "Show Compass Directions",
-  "feature.wynntils.map.overlay.miniMap.showCompass.name": "How should the cardinal directions (N/E/S/W) be displayed on the minimap?",
-  "feature.wynntils.map.overlay.miniMap.showCoords.description": "Should your coordinates be displayed below the minimap?",
-  "feature.wynntils.map.overlay.miniMap.showCoords.name": "Minimap Coordinates",
+  "feature.wynntils.map.pointerColor.description": "What colour should the main map pointer be?",
+  "feature.wynntils.map.pointerColor.name": "Pointer Colour",
+  "feature.wynntils.map.pointerType.description": "What style should the main map pointer have?",
+  "feature.wynntils.map.pointerType.name": "Main Map Pointer Type",
+  "feature.wynntils.miniMap.name": "Mini Map",
+  "feature.wynntils.miniMap.overlay.miniMap.borderType.description": "What style should the minimap border have?",
+  "feature.wynntils.miniMap.overlay.miniMap.borderType.name": "Mini Map Border Type",
+  "feature.wynntils.miniMap.overlay.miniMap.description": "Mini Map that displays region surrounding player",
+  "feature.wynntils.miniMap.overlay.miniMap.followPlayerRotation.description": "Should the minimap be locked facing north or rotate based on the direction you're facing?",
+  "feature.wynntils.miniMap.overlay.miniMap.followPlayerRotation.name": "Follow Player Rotation",
+  "feature.wynntils.miniMap.overlay.miniMap.maskType.description": "Should the minimap be a rectangle or a circle? (Circle is unsupported as of now)?",
+  "feature.wynntils.miniMap.overlay.miniMap.maskType.name": "Map Mask Type",
+  "feature.wynntils.miniMap.overlay.miniMap.name": "Mini Map Overlay",
+  "feature.wynntils.miniMap.overlay.miniMap.pointerColor.description": "What colour should the minimap pointer be?",
+  "feature.wynntils.miniMap.overlay.miniMap.pointerColor.name": "Pointer Colour",
+  "feature.wynntils.miniMap.overlay.miniMap.pointerType.description": "What style should the minimap pointer have?",
+  "feature.wynntils.miniMap.overlay.miniMap.pointerType.name": "Mini Map Pointer Type",
+  "feature.wynntils.miniMap.overlay.miniMap.renderUsingLinear.description": "Should the map be rendered using linear textures to avoid aliasing issues?",
+  "feature.wynntils.miniMap.overlay.miniMap.renderUsingLinear.name": "Use Linear Rendering",
+  "feature.wynntils.miniMap.overlay.miniMap.scale.description": "How zoomed in should the map be?",
+  "feature.wynntils.miniMap.overlay.miniMap.scale.name": "Map Scale",
+  "feature.wynntils.miniMap.overlay.miniMap.showCompass.description": "Show Compass Directions",
+  "feature.wynntils.miniMap.overlay.miniMap.showCompass.name": "How should the cardinal directions (N/E/S/W) be displayed on the minimap?",
+  "feature.wynntils.miniMap.overlay.miniMap.showCoords.description": "Should your coordinates be displayed below the minimap?",
+  "feature.wynntils.miniMap.overlay.miniMap.showCoords.name": "Minimap Coordinates",
   "feature.wynntils.mountHorseHotkey.alreadyRiding": "You are already riding a horse.",
   "feature.wynntils.mountHorseHotkey.name": "Mount Horse Hotkey",
   "feature.wynntils.mountHorseHotkey.noHorse": "Your horse was unable to be found.",
